Move download logic out of API controller
