+++ pkg/font/installer.go
@@ -4,6 +4,8 @@ import (
 	"fmt"
 	"fontman/client/pkg/api"
 	"fontman/client/pkg/util"
+	"io/ioutil"
+	"net/http"
 	"os"
 	"path/filepath"
 )
@@ -20,6 +22,25 @@ func (i InstallationError) Error() string {
 func validateFormat(file string) bool {
 	fileType := filepath.Ext(file)
 	return fileType == ".ttf" || fileType == ".otf" || fileType == ".ttc"
+}
+
+// DownloadFrom: downloads file from 'url' and saves it as 'dest`
+func DownloadFrom(url string, dest string) error {
+	response, resErr := http.Get(url)
+
+	if resErr != nil {
+		return resErr
+	}
+
+	defer response.Body.Close()
+
+	contents, readErr := ioutil.ReadAll(response.Body)
+
+	if readErr != nil {
+		return readErr
+	}
+
+	return os.WriteFile(dest, contents, 0777)
 }
 
 // InstallFont install a font either locally or globally & regenerate the cache.
@@ -70,13 +91,20 @@ func InstallFont(file string, isGlobal bool) error {
 }
 
 func InstallFromRemote(id string) error {
-	remoteFont, remoteErr := api.GetFontDetails(id)
+	font, err := api.GetFontDetails(id)
 
-	if remoteErr != nil {
-		return remoteErr
+	if err != nil {
+		return err
 	}
 
-	fmt.Println("downloaded:", remoteFont)
+	// download each style to a file with name: <family>-<style>.<format>
+	for _, style := range font.Styles {
+		dest := fmt.Sprintf("%s-%s.%s", font.Name, style.Type, "ttf")
+
+		if err := DownloadFrom(style.Url, dest); err != nil {
+			return err
+		}
+	}
 
 	return nil
 }
